* crash fixes + Unit::HasAura * optimizations
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@2251 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/game/LootMgr.h
@@ -55,7 +55,7 @@ typedef struct
 	uint32 displayid;
 }_LootItem;
 
-typedef std::set<uint64> LooterSet;
+typedef std::set<uint32> LooterSet;
 
 typedef struct
 {
